+++ cac_jira/commands/issue/delete.py
@@ -1,9 +1,10 @@
 # #!/usr/bin/env python
-# pylint: disable=no-member
+# pylint: disable=broad-exception-caught
 
-from jira_cmd.commands.issue import JiraIssueCommand
+from cac_jira.commands.issue import JiraIssueCommand
 
-class IssueComment(JiraIssueCommand):
+
+class IssueDelete(JiraIssueCommand):
     """
     Command class for commenting on Jira issues.
     """
@@ -24,13 +25,6 @@ class IssueComment(JiraIssueCommand):
             default=None,
             required=True,
         )
-        parser.add_argument(
-            "-c",
-            "--comment",
-            help="Comment to add",
-            default=None,
-            required=True,
-        )
         return parser
 
     def execute(self, args):
@@ -40,6 +34,10 @@ class IssueComment(JiraIssueCommand):
         Args:
             args: The parsed arguments
         """
-        self.log.debug("Commenting on Jira issue")
-        self.jira_client.add_comment(args.issue, args.comment)
-        self.log.info("Added comment to %s", args.issue)
+        self.log.debug("Deleting Jira issue")
+        try:
+            self.jira_client.delete_issue(args.issue)
+        except Exception as e:
+            self.log.error("Failed to find issue %s: %s", args.issue, e)
+            return
+        self.log.info("Issue %s deleted", args.issue)
