+++ cac_jira/commands/project/list.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import cac_core as cac
-from jira_cmd.commands.project import JiraProjectCommand
+from cac_jira.commands.project import JiraProjectCommand
 
 
 class ProjectList(JiraProjectCommand):
