+++ cac_jira/commands/project/show.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
 
 import cac_core as cac
-from jira_cmd.commands.project import JiraProjectCommand
-from jira_cmd.commands.project.list import ProjectList
+from cac_jira.commands.project import JiraProjectCommand
+from cac_jira.commands.project.list import ProjectList
 
 class ProjectShow(JiraProjectCommand):
     """
